Move local env vars to microservices
