+++ microservices/chat-agents-executor/chat_agents_executor/state_handlers/ask_gemini_handler.py
@@ -1,5 +1,5 @@
 from typing import Union, Optional
-from persona_sync_pylib.queue import publish_message
+from persona_sync_pylib.queue import publish_chat_agents_message
 from persona_sync_pylib.utils.singleton import singleton
 from persona_sync_pylib.types.chat_agents import (
     QueueRequest,
@@ -9,7 +9,8 @@ from persona_sync_pylib.types.chat_agents import (
 )
 from persona_sync_pylib.utils.logger import Logger, LogLevel
 
-from ..gemini import GeminiAPIDao
+from ..utils.gemini import GeminiAPIDao
+from ..utils.environment import QUEUE_NAME
 from .handler import Handler
 
 
@@ -77,8 +78,12 @@ Do not make up additional questions. The number of <ASK> and <ANS> tags should b
             transition_request.q_and_a_s = questions
 
             transition_request.state = PromptState.ASK_USER
-            publish_message(message=transition_request)
+            publish_chat_agents_message(
+                message=transition_request, queue_name=QUEUE_NAME
+            )
         else:
             transition_request.q_and_a_s = answered_q_and_a_s
             transition_request.state = PromptState.COMMUNICATE
-            publish_message(message=transition_request)
+            publish_chat_agents_message(
+                message=transition_request, queue_name=QUEUE_NAME
+            )
