+++ microservices/chat-agents-executor/chat_agents_executor/state_handlers/communication_handler.py
@@ -1,5 +1,5 @@
 from typing import Union, Optional
-from persona_sync_pylib.queue import publish_message
+from persona_sync_pylib.queue import publish_chat_agents_message
 from persona_sync_pylib.utils.singleton import singleton
 from persona_sync_pylib.types.chat_agents import (
     QueueRequest,
@@ -8,9 +8,9 @@ from persona_sync_pylib.types.chat_agents import (
     QAndA,
 )
 from persona_sync_pylib.utils.logger import Logger, LogLevel
-from persona_sync_pylib.utils.environment import MAX_INTERACTIONS
 
-from ..gemini import GeminiAPIDao
+from ..utils.environment import MAX_INTERACTIONS, QUEUE_NAME
+from ..utils.gemini import GeminiAPIDao
 from .handler import Handler
 
 
@@ -85,7 +85,9 @@ generate that.
         if model_output == "TO TERMINAL":
             transition_request = prompt_request.model_copy(deep=True)
             transition_request.state = PromptState.TERMINAL
-            publish_message(message=transition_request)
+            publish_chat_agents_message(
+                message=transition_request, queue_name=QUEUE_NAME
+            )
             return
 
         questions = super()._extract_text_between_tags(model_output, "ASK")
@@ -99,9 +101,13 @@ generate that.
             ]
             transition_request.previous_response = model_output
             transition_request.state = PromptState.ASK_GEMINI
-            publish_message(message=transition_request)
+            publish_chat_agents_message(
+                message=transition_request, queue_name=QUEUE_NAME
+            )
         else:
             transition_request.target = "u1" if prompt_request.target == "u2" else "u2"
             transition_request.previous_response = model_output
             transition_request.state = PromptState.COMMUNICATE
-            publish_message(message=transition_request)
+            publish_chat_agents_message(
+                message=transition_request, queue_name=QUEUE_NAME
+            )
