+++ microservices/chat-agents-executor/chat_agents_executor/utils/environment.py
@@ -1,0 +1,11 @@
+import os
+
+
+QUEUE_NAME = os.getenv("GEMINI_MESSAGING_QUEUE_NAME", "gemini-messaging")
+
+MONGO_DB_NAME = os.getenv("MONGO_DB_NAME", "chat-agents")
+PROMPT_INPUTS_COLLECTION = os.getenv("PROMPT_INPUTS_COLLECTION", "prompt-inputs")
+
+MAX_INTERACTIONS = os.getenv("MAX_INTERACTIONS", 4)
+
+GEMINI_MODEL = os.getenv("GEMINI_MODEL", "gemini-1.5-flash")
