+++ microservices/gemini-agents-service/app/model/initiate_request.py
@@ -1,8 +1,6 @@
 from pydantic import BaseModel
 from persona_sync_pylib.types.chat_agents import StateMachineQueueRequest
 
-from app.model.prompt_request import PromptRabbitRequest
-
 
 class InitiateChatHTTPRequest(BaseModel):
     u1_uid: str
