+++ microservices/gemini-agents-service/app/routers/gemini_agents.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter
 from fastapi.exceptions import HTTPException
-from persona_sync_pylib.queue import publish_message
+from persona_sync_pylib.queue import publish_chat_agents_message
 
 from app.store.prompt_input_dao import PromptInputDao
 from app.model.prompt_request import PromptHTTPRequest, PromptRabbitRequest
@@ -9,6 +9,7 @@ from app.model.initiate_request import (
     InitiateChatRabbitRequest,
 )
 from app.model.revert_request import RevertChatHTTPRequest, RevertChatRabbitRequest
+from app.utils.environment import QUEUE_NAME
 
 
 router = APIRouter()
@@ -19,7 +20,9 @@ async def prompt(prompt_http_request: PromptHTTPRequest):
     rabbit_request = PromptRabbitRequest(
         input=prompt_http_request.prompt, state="PROMPT"
     )
-    publish_status = publish_message(message=rabbit_request)
+    publish_status = publish_chat_agents_message(
+        message=rabbit_request, queue_name=QUEUE_NAME
+    )
 
     return {"status": publish_status}
 
@@ -35,7 +38,9 @@ async def chat(initiate_chat_http_request: InitiateChatHTTPRequest):
         u2_summary=initiate_chat_http_request.u2_summary,
         target="u1",
     )
-    publish_status = publish_message(message=rabbit_request)
+    publish_status = publish_chat_agents_message(
+        message=rabbit_request, queue_name=QUEUE_NAME
+    )
 
     return {"status": publish_status}
 
@@ -56,7 +61,9 @@ async def chat_revert(
 
     try:
         rabbit_request = RevertChatRabbitRequest(**result)
-        publish_status = publish_message(message=rabbit_request)
+        publish_status = publish_chat_agents_message(
+            message=rabbit_request, queue_name=QUEUE_NAME
+        )
 
         return {"status": publish_status}
     except Exception as e:
